refactor: relocate legacy facade modules to adapters layer (ARCHITECTURE §4.2, Resolves #291)
PR #290 placed legacy facades under app/<domain>/application/legacy.py,
but those modules import sqlalchemy.orm.Session and SqlAlchemy*UnitOfWork
adapters. ARCHITECTURE §4.2 forbids application/ from depending on
adapters/, so move them into adapters/ alongside the precedent set by
app/servers/adapters/_legacy_helpers.py (PR #275).

Moved 3 files (rename, content unchanged):
- app/backups/application/legacy.py -> app/backups/adapters/legacy.py
- app/files/application/legacy.py   -> app/files/adapters/legacy.py
- app/templates/application/legacy.py -> app/templates/adapters/legacy.py

Updated 5 callsites (1 production, 4 tests):
- app/files/application/management.py
- tests/test_security.py (3 references, incl. patch target string)
- tests/unit/backups/test_legacy_shim.py (4 imports)
- tests/unit/files/test_legacy_shim.py
- tests/unit/templates/test_legacy_shim.py

app/groups/application/legacy.py is intentionally left in place: its
_LegacyGroupFacade has no Session/adapter dependency and is
ARCHITECTURE-compliant in application/.
