+++ app/files/adapters/legacy.py
