+++ app/templates/adapters/legacy.py
